Move sniffForPath to a more better place
